Added background and hotkey program
Now that I'm on NixOS I would like to stop using arbitrary scripts
like .xinitrc and what not so I always get a consistent build.
Of course there is a long way yet to go, since my XMonad config
depends on a range of stuff that is not explicit, but I'm getting there.