fix(xmonad): make mod+a always switch to first screen from left
By using XMonad.Actions.PhysicalScreens we get screen numbering that
counts from left to right. This means that keybindings always switch
to the same screen, i.e. mod+a always switches to the first screen
from the left and mod+s always switches to the second screen from the
left.