fix(emacs): advice server-start to fix permissions
On NixOS the /run/user/$UID/emacs directory has the wrong
permission (u+rx and not u+rwx) and after failing to figure out why
this is I gave up and instead added my first ever Emacs advice.

The advice tells `server-start` to check that the `server-socket-dir`
has the correct owner permissions and otherwise fixes it.