fix(emacs): handle inner projects better
When we're working with projects that are not at the root of a version
control repositories the standard project package has some troubles
locating what it should consider a project.

Previously we tried to use buffer-local hooks, but that got tricky
when visiting other fiels, because the hook did not always load (a
little unclear why).

Instead this time we define functions that always run, but they exit
early with nil if they notice that the mode is not what they expect.

This whole machinery is hidden by some helper functions, i.e.:

    ;; a Go module is a directory containing a go.mod file
    (find-inner-project "go" (lambda (dir) (locate-dominating-file dir "go.mod")))

Since we need to be able to capture the "go" value, we also needed to
activate lexical-binding in our init.el.